server/requirements.txt
Flask==2.0.2
Werkzeug==2.3.7
flask-cors==4.0.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
ruff>=0.1.0